transfer PDFReader to ImgReader which is more common in usage
