+++ contributors/CNSeniorious000/pdf.py
@@ -67,46 +67,3 @@ class PDF:
 
         for page in it:
             self.save_image(f"{dir_path}/{page}.{encode}", page, dpi, alpha)
-
-
-# PDF与OCR
-class PDFReader(PDF):
-    def __init__(self, path: str, lang=("en", "ch_sim")):
-        super().__init__(path)
-        self.lang = lang
-
-    @cached_property
-    def reader(self):
-        from easyocr import Reader
-        return Reader(self.lang)
-
-    @cache
-    def get_texts(self, page=0, *args, **kwargs):
-        return self.reader.readtext(self.get_image(page, *args, **kwargs))
-
-    def render_bbox(self, page=0, *args, color=(255, 0, 0), **kwargs):
-        from cv2 import line
-        image = self.get_image(page, *args, **kwargs).copy()
-        for points, string, degree in self.get_texts(page, *args, **kwargs):
-            c = (*color, round(degree * 255))
-            # noinspection PyPep8Naming
-            A, B, C, D = [tuple(map(round, point)) for point in points]
-            line(image, A, B, c)
-            line(image, B, C, c)
-            line(image, C, D, c)
-            line(image, D, A, c)
-
-        return image
-
-    def show_image_with_bbox(self, *args, **kwargs):
-        return self.show(self.render_bbox(*args, **kwargs))
-
-    def classify(self, choices, page=0, *args, **kwargs):
-        """
-        这个函数的功能在于，比如有一个奖状，上面可能有一等奖二等奖三等奖，要批量识别出一张图片为什么奖
-        :param page: the page to ocr
-        :param choices: list or words to choose
-        """
-        from rapidfuzz.process import extractOne
-        texts = [string for _, string, _ in self.get_texts(page, *args, **kwargs)]
-        return sorted((extractOne(choice, texts)[1], choice) for choice in choices)[-1][1]
